requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.11.1
boto3==1.25.3
botocore==1.28.3
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-static==0.0.6
Django==4.1.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-redis==5.2.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
et-xmlfile==1.1.0
idna==2.10
jmespath==1.0.1
lxml==4.9.2
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.2
parse==1.19.0
passlib==1.7.4
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.5
redis==4.4.2
requests==2.25.1
s3transfer==0.6.0
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
static3==0.7.0
tzdata==2022.5
urllib3==1.26.12
